Refactor tags and criteria into taggable
